Tidy up style, warnings, and added convenience support for Strings as Java var data.

+++ test/java/uk/co/real_logic/sbe/generation/java/JavaGeneratorTest.java
@@ -43,8 +43,8 @@ public class JavaGeneratorTest
 {
     private static final Class<?> BUFFER_IMPLEMENTATION = MutableDirectBuffer.class;
     private static final String BUFFER_IMPLEMENTATION_NAME = BUFFER_IMPLEMENTATION.getName();
-
     private static final ByteOrder BYTE_ORDER = ByteOrder.nativeOrder();
+
     private final StringWriterOutputManager outputManager = new StringWriterOutputManager();
     private final MutableDirectBuffer mockBuffer = mock(MutableDirectBuffer.class);
 
@@ -67,12 +67,12 @@ public class JavaGeneratorTest
     {
         final int bufferOffset = 64;
         final int templateIdOffset = 2;
-        final Short templateId = Short.valueOf((short)7);
+        final short templateId = (short)7;
         final int actingVersion = 0;
-        final Integer blockLength = Integer.valueOf(32);
+        final int blockLength = 32;
         final String fqClassName = ir.applicableNamespace() + "." + MESSAGE_HEADER_TYPE;
 
-        when(Short.valueOf(mockBuffer.getShort(bufferOffset + templateIdOffset, BYTE_ORDER))).thenReturn(templateId);
+        when(mockBuffer.getShort(bufferOffset + templateIdOffset, BYTE_ORDER)).thenReturn(templateId);
 
         final JavaGenerator javaGenerator = new JavaGenerator(ir, BUFFER_IMPLEMENTATION_NAME, outputManager);
         javaGenerator.generateMessageHeaderStub();
@@ -82,14 +82,14 @@ public class JavaGeneratorTest
 
         final Object flyweight = clazz.newInstance();
         final Method method = flyweight.getClass().getDeclaredMethod("wrap", BUFFER_IMPLEMENTATION, int.class, int.class);
-        method.invoke(flyweight, mockBuffer, Integer.valueOf(bufferOffset), Integer.valueOf(actingVersion));
+        method.invoke(flyweight, mockBuffer, bufferOffset, actingVersion);
 
         final Integer result = (Integer)clazz.getDeclaredMethod("templateId").invoke(flyweight);
-        assertThat(result, is(Integer.valueOf(templateId.intValue())));
+        assertThat(result, is((int)templateId));
 
         clazz.getDeclaredMethod("blockLength", int.class).invoke(flyweight, blockLength);
 
-        verify(mockBuffer).putShort(bufferOffset, blockLength.shortValue(), BYTE_ORDER);
+        verify(mockBuffer).putShort(bufferOffset, (short)blockLength, BYTE_ORDER);
     }
 
     @Test
@@ -104,7 +104,7 @@ public class JavaGeneratorTest
         final Class<?> clazz = CompilerUtil.compileInMemory(fqClassName, outputManager.getSources());
         assertNotNull(clazz);
 
-        final Object result = clazz.getDeclaredMethod("get", short.class).invoke(null, Short.valueOf((short)1));
+        final Object result = clazz.getDeclaredMethod("get", short.class).invoke(null, (short)1);
 
         assertThat(result.toString(), is("TRUE"));
     }
@@ -121,7 +121,7 @@ public class JavaGeneratorTest
         final Class<?> clazz = CompilerUtil.compileInMemory(fqClassName, outputManager.getSources());
         assertNotNull(clazz);
 
-        final Object result = clazz.getDeclaredMethod("get", byte.class).invoke(null, Byte.valueOf((byte)'B'));
+        final Object result = clazz.getDeclaredMethod("get", byte.class).invoke(null, (byte)'B');
 
         assertThat(result.toString(), is("B"));
     }
@@ -131,11 +131,11 @@ public class JavaGeneratorTest
     {
         final int bufferOffset = 8;
         final int actingVersion = 0;
-        final Byte bitset = Byte.valueOf((byte)0b0000_0100);
+        final byte bitset = (byte)0b0000_0100;
         final String className = "OptionalExtras";
         final String fqClassName = ir.applicableNamespace() + "." + className;
 
-        when(Byte.valueOf(mockBuffer.getByte(bufferOffset))).thenReturn(bitset);
+        when(mockBuffer.getByte(bufferOffset)).thenReturn(bitset);
 
         final JavaGenerator javaGenerator = new JavaGenerator(ir, BUFFER_IMPLEMENTATION_NAME, outputManager);
         javaGenerator.generateTypeStubs();
@@ -145,7 +145,7 @@ public class JavaGeneratorTest
 
         final Object flyweight = clazz.newInstance();
         final Method method = flyweight.getClass().getDeclaredMethod("wrap", BUFFER_IMPLEMENTATION, int.class, int.class);
-        method.invoke(flyweight, mockBuffer, Integer.valueOf(bufferOffset), Integer.valueOf(actingVersion));
+        method.invoke(flyweight, mockBuffer, bufferOffset, actingVersion);
 
         final Object result = clazz.getDeclaredMethod("cruiseControl").invoke(flyweight);
 
@@ -166,8 +166,8 @@ public class JavaGeneratorTest
         final String className = "Engine";
         final String fqClassName = ir.applicableNamespace() + "." + className;
 
-        when(Short.valueOf(mockBuffer.getShort(capacityFieldOffset, BYTE_ORDER)))
-            .thenReturn(Short.valueOf((short)expectedEngineCapacity));
+        when(mockBuffer.getShort(capacityFieldOffset, BYTE_ORDER))
+            .thenReturn((short)expectedEngineCapacity);
 
         final JavaGenerator javaGenerator = new JavaGenerator(ir, BUFFER_IMPLEMENTATION_NAME, outputManager);
         javaGenerator.generateTypeStubs();
@@ -177,19 +177,19 @@ public class JavaGeneratorTest
 
         final Object flyweight = clazz.newInstance();
         final Method method = flyweight.getClass().getDeclaredMethod("wrap", BUFFER_IMPLEMENTATION, int.class, int.class);
-        method.invoke(flyweight, mockBuffer, Integer.valueOf(bufferOffset), Integer.valueOf(actingVersion));
+        method.invoke(flyweight, mockBuffer, bufferOffset, actingVersion);
 
-        final Integer capacityResult = (Integer)clazz.getDeclaredMethod("capacity").invoke(flyweight);
-        assertThat(capacityResult, is(Integer.valueOf(expectedEngineCapacity)));
+        final int capacityResult = (Integer)clazz.getDeclaredMethod("capacity").invoke(flyweight);
+        assertThat(capacityResult, is(expectedEngineCapacity));
 
-        final Integer maxRpmResult = (Integer)clazz.getDeclaredMethod("maxRpm").invoke(flyweight);
-        assertThat(maxRpmResult, is(Integer.valueOf(expectedMaxRpm)));
+        final int maxRpmResult = (Integer)clazz.getDeclaredMethod("maxRpm").invoke(flyweight);
+        assertThat(maxRpmResult, is(expectedMaxRpm));
 
         final short numCylinders = (short)4;
-        clazz.getDeclaredMethod("numCylinders", short.class).invoke(flyweight, Short.valueOf(numCylinders));
+        clazz.getDeclaredMethod("numCylinders", short.class).invoke(flyweight, numCylinders);
 
         clazz.getDeclaredMethod("putManufacturerCode", byte[].class, int.class)
-             .invoke(flyweight, manufacturerCode, Integer.valueOf(0));
+            .invoke(flyweight, manufacturerCode, 0);
 
         verify(mockBuffer).putByte(numCylindersOffset, (byte)numCylinders);
         verify(mockBuffer).putBytes(manufacturerCodeOffset, manufacturerCode, 0, manufacturerCode.length);
@@ -209,8 +209,9 @@ public class JavaGeneratorTest
         assertNotNull(clazz);
 
         final Object msgFlyweight = clazz.newInstance();
-        msgFlyweight.getClass().getDeclaredMethod("wrapForEncode", BUFFER_IMPLEMENTATION, int.class)
-                               .invoke(msgFlyweight, buffer, Integer.valueOf(0));
+        msgFlyweight.getClass()
+            .getDeclaredMethod("wrapForEncode", BUFFER_IMPLEMENTATION, int.class)
+            .invoke(msgFlyweight, buffer, 0);
 
         final Integer initialPosition = (Integer)msgFlyweight.getClass().getDeclaredMethod("limit").invoke(msgFlyweight);
 
@@ -219,7 +220,7 @@ public class JavaGeneratorTest
             (Integer)msgFlyweight.getClass().getDeclaredMethod("limit").invoke(msgFlyweight), greaterThan(initialPosition));
 
         final Integer count = (Integer)groupFlyweight.getClass().getDeclaredMethod("count").invoke(groupFlyweight);
-        assertThat(count, is(Integer.valueOf(0)));
+        assertThat(count, is(0));
     }
 
     @Test(expected = IllegalArgumentException.class)
@@ -233,5 +234,4 @@ public class JavaGeneratorTest
     {
         new JavaGenerator(ir, "java.nio.ByteBuffer", outputManager);
     }
-
 }
